pyproject.toml
[project]
name = "Q-Trader"
version = "0.1.0"
description = "量化自动化交易程序"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tqsdk>=2.15.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "aiosqlite>=0.19.0",
    "loguru>=0.7.2",
    "pyyaml>=6.0.1",
    "pandas>=2.1.3",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "watchdog>=3.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
files = [
    "src/",
    "!src/__pycache__/",
]
